Optional data type handling changes

+++ HandsetDemoProject/src/main/java/com/example/dao/HandsetDetailsDao.java
@@ -2,10 +2,11 @@ package com.example.dao;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 
-import com.example.model.Handset;
+import com.example.model.Sensor;
 import com.example.model.Mobile;
 
-public interface HandsetDetailsDao extends JpaRepository<Mobile, Integer>{
+public interface HandsetDetailsDao extends JpaRepository<Mobile, Long>{
 	public Mobile findByModelName(String name);
 }
